use a spatial index in EmpiricalVariogram
If there are many points involved, and the search radius is small relative to the complete domain, using a spatial index is more efficient than iterating over half the points. For one real world dataset this sped up the algorithm about 20x, though of course it greatly varies. Results are identical.

Besides the spatial index, adding an early exit for if `h > hmax` already speeds up the code, avoiding the expensive `v` calculation.